feat: delete image in storage when deleting

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require('../models/product');
 const { validationResult } = require('express-validator');
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
@@ -144,27 +145,28 @@ exports.postEditProduct = (req, res, next) => {
     });
   }
 
-  Product.findById(prodId).then((product) => {
-    if (product.userId.toString() !== req.user._id.toString()) {
-      return res.redirect('/');
-    }
-    product.title = updatedTitle;
-    product.price = updatedPrice;
-    product.description = updatedDescription;
-    if (image) {
-      product.imageUrl = image.path;
-    }
-    return product.save().then((result) => {
-      console.log('UPDATED PRODUCT!');
-      res.redirect('/admin/products');
+  Product.findById(prodId)
+    .then((product) => {
+      if (product.userId.toString() !== req.user._id.toString()) {
+        return res.redirect('/');
+      }
+      product.title = updatedTitle;
+      product.price = updatedPrice;
+      product.description = updatedDescription;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
+      return product.save().then((result) => {
+        console.log('UPDATED PRODUCT!');
+        res.redirect('/admin/products');
+      });
+    })
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
     });
-  }).catch((err)=>{
-    const error = new Error(err)
-    error.httpStatusCode = 500
-    return next(error)
-
-  })
- 
 };
 
 exports.getAdminProducts = (req, res, next) => {
@@ -184,7 +186,14 @@ exports.getAdminProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error('Product not found'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log('Deleted');
       res.redirect('/admin/products');
